commit-15 contact-partner

+++ src/main/java/uz/resultme/service/PartnerService.java
@@ -8,54 +8,65 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import uz.resultme.entity.Partner;
 import uz.resultme.entity.Photo;
+import uz.resultme.exception.IllegalPhotoTypeException;
 import uz.resultme.payload.ApiResponse;
 import uz.resultme.payload.PartnerDTO;
 import uz.resultme.repository.PartnerRepository;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
-public class PartnerService {
+public class PartnerService
+{
     private final PartnerRepository partnerRepository;
     private final PhotoService photoService;
 
-    public ResponseEntity<PartnerDTO> addPartner(MultipartFile file,String url){
+    public ResponseEntity<ApiResponse<PartnerDTO>> addPartner(MultipartFile file, String url)
+    {
 
-        ApiResponse<Partner> response=new ApiResponse<>();
+        ApiResponse<PartnerDTO> response = new ApiResponse<>();
         ObjectMapper objectMapper = new ObjectMapper();
         if (file.getContentType() == null ||
-                !file.getContentType().equals("image/png")||!file.getContentType().equals("image/svg"))
+                !file.getContentType().equals("image/png") || !file.getContentType().equals("image/svg"))
         {
-            throw new RuntimeException("Unsupported content type: " + file.getContentType());
+            throw new IllegalPhotoTypeException("Unsupported image type :" + file.getContentType());
         }
 
         Photo photo = photoService.save(file);
-        Partner partner=new Partner(photo,url);
+        Partner partner = new Partner(photo, url);
         partnerRepository.save(partner);
 
-        return ResponseEntity.status(HttpStatus.CREATED).body(new PartnerDTO(partner));
+        response.setMessage("Created");
+        response.setData(new PartnerDTO(partner));
+
+        return ResponseEntity.status(HttpStatus.CREATED).body(response);
     }
 
-    public ResponseEntity<ApiResponse<Partner>> updatePartner(Long id, MultipartFile file, String url) {
-        ApiResponse<Partner> response=new ApiResponse<>();
-        if (!partnerRepository.existsById(id)){
+    public ResponseEntity<ApiResponse<Partner>> updatePartner(Long id, MultipartFile file, String url)
+    {
+        ApiResponse<Partner> response = new ApiResponse<>();
+        if (!partnerRepository.existsById(id))
+        {
             response.setMessage("Partner with id " + id + " does not exist");
             return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
         }
-        Partner newPartner=partnerRepository.findById(id).get();
-        Photo oldPhoto=newPartner.getPhoto();
+        Partner newPartner = partnerRepository.findById(id).get();
+        Photo oldPhoto = newPartner.getPhoto();
 
-        if (file==null|| file.isEmpty()){
+        if (file == null || file.isEmpty())
+        {
             newPartner.setPhoto(oldPhoto);
-        }else {
+        } else
+        {
             newPartner.setPhoto(photoService.save(file));
         }
-        if(url==null||url.isEmpty()){
+        if (url == null || url.isEmpty())
+        {
             newPartner.setPartnerUrl(newPartner.getPartnerUrl());
-        }else {
+        } else
+        {
             newPartner.setPartnerUrl(url);
         }
         newPartner.setId(id);
@@ -73,7 +84,7 @@ public class PartnerService {
         ApiResponse<PartnerDTO> response = new ApiResponse<>();
         Partner partner = partnerRepository.findPartnerByPartnerUrl(url);
         response.setMessage("Found partner with url " + url);
-        return new ResponseEntity<>(response,HttpStatus.OK);
+        return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
     public ResponseEntity<ApiResponse<List<PartnerDTO>>> findAll()
@@ -90,6 +101,13 @@ public class PartnerService {
     public ResponseEntity<ApiResponse<Partner>> delete(Long id)
     {
         ApiResponse<Partner> response = new ApiResponse<>();
+
+        if (!partnerRepository.existsById(id))
+        {
+            response.setMessage("Partner with id " + id + " does not exist");
+            return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+        }
+
         try
         {
             partnerRepository.deleteById(id);
